Fix edge cases and add new test cases for subsort
The original solution from the CTCI book had a void return type for the
solution function; returning the index pair is far more useful in
practice (both in larger systems and for testing), so we will continue
with this approach.

These changes change the following behaviors of the solution:
* Return more appropriate indeces for arrays of length 0 or 1
* Return more appropriate indeces for arrays that are already sorted
* Add additional test cases for robustness

NOTE: in this implementation, "more appropriate indeces" have been
chosen to be -1. This would indicate to the client that no action needs
to be taken since an index of -1 is not a legal index for a primitive
array, and prevents the need for the client to perform null checking.

Related to #84