Clean up memory allocation for sum-lists solutions
Because garbage collection isn't a thing in C++ and, for the sake of
completion, we should be aware of the implications of memory allocation
in non-garbage-collected languages, the implementation files have been
updated with the correct delete calls.

For clarity, the solutions file is being left as is, because the
delete calls get splattered into recursive functions for the bonus
part of the problem.