Add writeup for first problem (Stacks + Queues) (#101)
Add a simplified version of the circular buffer problem from LeetCode;
I got this problem on an interview a while back and enjoyed the
conversation I had with my interviewer about it. It's much more
open-ended than most of our standard questions and will be a good warmup
for folks.